Increase MaterialTableSize 8 times
Now that we prefetch in material hash table we
can increase its size and gain something.

Hit rate is now of 98% from 92%

Speedup of 0.8%

No functional change.

+++ src/material.cpp
@@ -32,7 +32,7 @@ namespace {
   const Value EndgameLimit = Value(3998);
 
   // Scale factors used when one side has no more pawns
-  const int NoPawnsSF[4] = { 6, 12, 32 };
+  const uint8_t NoPawnsSF[4] = { 6, 12, 32 };
 
   // Polynomial material balance parameters
   const Value RedundantQueenPenalty = Value(320);
